refact(api): merge terminal_models.py into schemas_terminal.py (#5996)
Eliminates dedicated model file in favour of domain schema file.
Updated 5 importers: terminal.py, terminal_handlers.py, terminal_tools.py,
terminal_websocket/security.py, terminal_websocket/audit.py.
