+++ autobot-backend/api/terminal_handlers.py
@@ -31,7 +31,7 @@ from typing import Awaitable, Callable, Dict, Optional
 from fastapi import WebSocket
 
 # Import models from dedicated module (Issue #185)
-from api.terminal_models import (
+from api.schemas_terminal import (
     MODERATE_RISK_PATTERNS,
     RISKY_COMMAND_PATTERNS,
     CommandRiskLevel,
